refactor: разделение функционала по модулям
