+++ client/composables/index.ts
@@ -2,6 +2,8 @@ export * from './apollo-helpers'
 export * from './debounce'
 export * from './filters'
 export * from './grid'
+export * from './grid-mutations'
+export * from './grid-resizing'
 export * from './i18n'
 export * from './overlying-class'
 export * from './page'
